further split
